# api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "neit Api Documentation",
    "description": "neit Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "neit Logo",
      "altText": "neit Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "neit API Support",
      "url": "https://neit.com/info/support",
      "email": "support@neit.com"
    },
    "termsOfService": "https://neit.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/servicios": {
      "get": {
        "summary": "List Servicios",
        "description": "Servicios",
        "operationId": "ListServicios",
        
        "tags": [
          "Servicios"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListServicios"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/servicios/view/{id}": {
      "get": {
        "summary": "View Servicios",
        "description": "Servicios Details",
        "operationId": "ViewServicios",
        
        "tags": [
          "Servicios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewServicios"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/servicios/add": {
      "post": {
        "summary": "Add Servicios",
        "description": "Add New Servicios",
        "operationId": "AddServicios",
        
        "tags": [
          "Servicios"
        ],
        
        "requestBody": {
          "description": "Add Servicios",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddServicios"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewServicios"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/servicios/edit/{id}": {
      "put": {
        "summary": "Edit Servicios",
        "description": "Edit Servicios",
        "operationId": "EditServicios",
        
        "tags": [
          "Servicios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "requestBody": {
          "description": "Edit Servicios",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditServicios"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditServicios"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/servicios/delete/{id}": {
      "delete": {
        "summary": "Delete Servicios",
        "description": "Servicios",
        "operationId": "DeleteServicios",
        
        "tags": [
          "Servicios"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ListServicios": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "titulo": {
            "description": "Titulo",
            "type": "string"
          },
          "descripcion": {
            "description": "Descripcion",
            "type": "string"
          },
          "imagen": {
            "description": "Imagen",
            "type": "string"
          },
          "icono": {
            "description": "Icono",
            "type": "string"
          }
        }
      },
      "ViewServicios": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "titulo": {
            "description": "Titulo",
            "type": "string"
          },
          "descripcion": {
            "description": "Descripcion",
            "type": "string"
          },
          "imagen": {
            "description": "Imagen",
            "type": "string"
          },
          "icono": {
            "description": "Icono",
            "type": "string"
          }
        }
      },
      "AddServicios": {
        "type": "object",
        "properties": {
          "titulo": {
            "description": "Titulo",
            "type": "string"
          },
          "descripcion": {
            "description": "Descripcion",
            "type": "string"
          },
          "imagen": {
            "description": "Imagen",
            "type": "string"
          },
          "icono": {
            "description": "Icono",
            "type": "string"
          }
        }
      },
      "EditServicios": {
        "type": "object",
        "properties": {
          "titulo": {
            "description": "Titulo",
            "type": "string"
          },
          "descripcion": {
            "description": "Descripcion",
            "type": "string"
          },
          "imagen": {
            "description": "Imagen",
            "type": "string"
          },
          "icono": {
            "description": "Icono",
            "type": "string"
          }
        }
      },
...
      "SearchQuery": {
        "type": "string",
        "properties": {
          "search": {
            "description": "Search value",
            "type": "string"
          }
        }
      },
...
      "LimitQuery": {
        "type": "string",
        "properties": {
          "limit": {
            "description": "Maximum number of records to select",
            "type": "integer"
          }
        }
      },
      "PageNumberQuery": {
        "type": "string",
        "properties": {
          "pagenumber": {
            "description": "Current page number",
            "type": "integer"
          }
        }
      },
      "RecordID": {
        "type": "string",
        "properties": {
          "id": {
            "description": "Record ID",
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}