# nodejs-express-api/controllers/servicios.js
const express = require('express');
const router = express.Router();
const utils = require('../helpers/utils');
const { body, validationResult, matchedData } = require('express-validator');
const DB = require('../models/index.js');
const Servicios = DB.Servicios;


const sequelize = DB.sequelize; // sequelize query functions
const Op = DB.Op; // sequelize query operators
const filterBy = DB.filterBy; // sequelize where condtion




/**
 * Route to list servicios records
 * @route {GET} /servicios/index/{fieldname}/{fieldvalue}
 */
router.get(['/', '/index/:fieldname?/:fieldvalue?'], async (req, res) => {  
	try{
		let query = {};  // sequelize query object
		let queryFilters = [];
		let where = {};  // sequelize where conditions
		let replacements = {};  // sequelize query params
		let fieldName = req.params.fieldname;
		let fieldValue = req.params.fieldvalue;
		
		if (fieldName){
			queryFilters.push(filterBy(fieldName, fieldValue));
		}
		let search = req.query.search;
		if(search){
			let searchFields = Servicios.searchFields();
			where[Op.or] = searchFields;
			replacements.search = `%${search}%`;
		}
		
		if(queryFilters.length){
			where[Op.and] = queryFilters;
		}
		query.raw = true;
		query.where = where;
		query.replacements = replacements;
		query.order = Servicios.getOrderBy(req, 'desc');
		query.attributes = Servicios.listFields();
		let page = parseInt(req.query.page) || 1;
		let limit = parseInt(req.query.limit) || 10;
		let result = await Servicios.paginate(query, page, limit);
		return res.ok(result);
	}
	catch(err) {
		return res.serverError(err);
	}
});


/**
 * Route to view Servicios record
 * @route {GET} /servicios/view/{recid}
 */
router.get(['/view/:recid'], async (req, res) => {
	try{
		const recid = req.params.recid || null;
		const query = {}
		const where = {}
		where['id'] = recid;
		query.raw = true;
		query.where = where;
		query.attributes = Servicios.viewFields();
		let record = await Servicios.findOne(query);
		if(!record){
			return res.notFound();
		}
		return res.ok(record);
	}
	catch(err){
		return res.serverError(err);
	}
});


/**
 * Route to insert Servicios record
 * @route {POST} /servicios/add
 */
router.post('/add/' , 
	[
		body('titulo').optional({nullable: true, checkFalsy: true}),
		body('descripcion').optional({nullable: true, checkFalsy: true}),
		body('imagen').optional({nullable: true, checkFalsy: true}),
		body('icono').optional({nullable: true, checkFalsy: true}),
	]
, async function (req, res) {
	try{
		let errors = validationResult(req); // get validation errors if any
		if (!errors.isEmpty()) {
			let errorMsg = utils.formatValidationError(errors.array());
			return res.badRequest(errorMsg);
		}
		let modeldata = matchedData(req, { locations: ['body'] }); // get validated data
		
        // move uploaded file from temp directory to destination directory
		if(modeldata.imagen !== undefined) {
			let fileInfo = utils.moveUploadedFiles(modeldata.imagen, "imagen");
			modeldata.imagen = fileInfo.filepath;
		}
		
        // move uploaded file from temp directory to destination directory
		if(modeldata.icono !== undefined) {
			let fileInfo = utils.moveUploadedFiles(modeldata.icono, "icono");
			modeldata.icono = fileInfo.filepath;
		}
		
		//save Servicios record
		let record = await Servicios.create(modeldata);
		//await record.reload(); //reload the record from database
		const recid =  record['id'];
		
		return res.ok(record);
	} catch(err){
		return res.serverError(err);
	}
});


/**
 * Route to get  Servicios record for edit
 * @route {GET} /servicios/edit/{recid}
 */
router.get('/edit/:recid', async (req, res) => {
	try{
		const recid = req.params.recid;
		const query = {};
		const where = {};
		where['id'] = recid;
		query.raw = true;
		query.where = where;
		query.attributes = Servicios.editFields();
		let record = await Servicios.findOne(query);
		if(!record){
			return res.notFound();
		}
		return res.ok(record);
	}
	catch(err){
		return res.serverError(err);
	}
});


/**
 * Route to update  Servicios record
 * @route {POST} /servicios/edit/{recid}
 */
router.post('/edit/:recid' , 
	[
		body('titulo').optional({nullable: true, checkFalsy: true}),
		body('descripcion').optional({nullable: true, checkFalsy: true}),
		body('imagen').optional({nullable: true, checkFalsy: true}),
		body('icono').optional({nullable: true, checkFalsy: true}),
	]
, async (req, res) => {
	try{
		let errors = validationResult(req); // get validation errors if any
		if (!errors.isEmpty()) {
			let errorMsg = utils.formatValidationError(errors.array());
			return res.badRequest(errorMsg);
		}
		const recid = req.params.recid;
		let modeldata = matchedData(req, { locations: ['body'], includeOptionals: true }); // get validated data
		
        // move uploaded file from temp directory to destination directory
		if(modeldata.imagen !== undefined) {
			let fileInfo = utils.moveUploadedFiles(modeldata.imagen, "imagen");
			modeldata.imagen = fileInfo.filepath;
		}
		
        // move uploaded file from temp directory to destination directory
		if(modeldata.icono !== undefined) {
			let fileInfo = utils.moveUploadedFiles(modeldata.icono, "icono");
			modeldata.icono = fileInfo.filepath;
		}
		const query = {};
		const where = {};
		where['id'] = recid;
		query.raw = true;
		query.where = where;
		query.attributes = Servicios.editFields();
		let record = await Servicios.findOne(query);
		if(!record){
			return res.notFound();
		}
		await Servicios.update(modeldata, {where: where});
		return res.ok(modeldata);
	}
	catch(err){
		return res.serverError(err);
	}
});


/**
 * Route to delete Servicios record by table primary key
 * Multi delete supported by recid separated by comma(,)
 * @route {GET} /servicios/delete/{recid}
 */
router.get('/delete/:recid', async (req, res) => {
	try{
		const recid = (req.params.recid || '').split(',');
		const query = {};
		const where = {};
		where['id'] = recid;
		query.raw = true;
		query.where = where;
		let records = await Servicios.findAll(query);
		records.forEach(async (record) => { 
			//perform action on each record before delete
		});
		await Servicios.destroy(query);
		return res.ok(recid);
	}
	catch(err){
		return res.serverError(err);
	}
});
module.exports = router;